Добавлена подержка Liquibase, прописано создание таблиц БД, добавлены сервисы, маперы
