+++ src/main/java/ru/crmkrd/projectmanager/entity/History.java
@@ -1,20 +1,23 @@
 package ru.crmkrd.projectmanager.entity;
 
+import io.swagger.v3.oas.annotations.Hidden;
+
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Id;
 import java.time.LocalDate;
 
 @Entity
+@Hidden
 public class History {
     @Id
     @Column(name = "id", nullable = false)
     private Long id;
 
-    @Column(name = "Description", length = 180)
+    @Column(name = "description", length = 180)
     private String description;
 
-    @Column(name = "DateTime")
+    @Column(name = "date_time")
     private LocalDate dateTime;
 
     public LocalDate getDateTime() {
