+++ src/main/java/ru/crmkrd/projectmanager/entity/Role.java
@@ -1,16 +1,19 @@
 package ru.crmkrd.projectmanager.entity;
 
+import io.swagger.v3.oas.annotations.Hidden;
+
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Id;
 
 @Entity
+@Hidden
 public class Role {
     @Id
     @Column(name = "id", nullable = false)
     private Long id;
 
-    @Column(name = "Role", length = 10)
+    @Column(name = "role", length = 10)
     private String role;
 
     public String getRole() {
