+++ src/main/java/ru/crmkrd/projectmanager/entity/User.java
@@ -1,22 +1,25 @@
 package ru.crmkrd.projectmanager.entity;
 
+import io.swagger.v3.oas.annotations.Hidden;
+
 import javax.persistence.*;
 
 @Table(name = "Users")
 @Entity
+@Hidden
 public class User {
     @Id
     @Column(name = "id", nullable = false)
     private Long id;
 
-    @Column(name = "Name", length = 30)
+    @Column(name = "name", length = 30)
     private String name;
 
-    @Column(name = "SecondName", length = 30)
+    @Column(name = "second_name", length = 30)
     private String secondName;
 
     @ManyToOne
-    @JoinColumn(name = "Role")
+    @JoinColumn(name = "role")
     private Role role;
 
     public Role getRole() {
