+++ src/main/resources/application.properties
@@ -1,11 +1,14 @@
 
-spring.datasource.url=jdbc:postgresql://localhost:5000/pmbase
-spring.datasource.username=postgres
-spring.datasource.password=222324
+spring.datasource.url = jdbc:postgresql://localhost:5000/pmbase
+spring.datasource.username = postgres
+spring.datasource.password = postgres
 
-spring.jpa.show-sql=true
-spring.jpa.properties.hibernate.format_sql=true
-spring.jpa.generate-ddl=false
-spring.jpa.hibernate.ddl-auto=none
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
-spring.jpa.properties.hibernate.jdbc.lob.non_contextual_creation=true
+spring.jpa.show-sql = true
+spring.jpa.properties.hibernate.format_sql = true
+spring.jpa.generate-ddl = false
+spring.jpa.hibernate.ddl-auto = none
+spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.PostgreSQLDialect
+spring.jpa.properties.hibernate.jdbc.lob.non_contextual_creation = true
+
+spring.liquibase.enabled = true
+spring.liquibase.change-log = db/changelog/db.changelog-master.yaml
