+++ src/main/resources/db/changelog/changeset/create-db-table.xml
@@ -1,0 +1,90 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<databaseChangeLog
+        xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
+        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog
+                      http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-4.3.xsd"
+        objectQuotingStrategy="QUOTE_ONLY_RESERVED_WORDS">
+    <changeSet
+            id="create-data-table"
+            author="Evtefeev">
+        <createTable tableName="history">
+            <column name="id" type="BIGINT">
+                <constraints nullable="false" primaryKey="true" primaryKeyName="pk_history"/>
+            </column>
+            <column name="description" type="VARCHAR(180)"/>
+            <column name="date_time" type="DATE"/>
+        </createTable>
+        <createTable tableName="project">
+            <column name="id" type="BIGINT">
+                <constraints nullable="false" primaryKey="true" primaryKeyName="pk_project"/>
+            </column>
+            <column name="project_name" type="VARCHAR(30)"/>
+            <column name="description" type="VARCHAR(120)"/>
+            <column name="autor" type="BIGINT"/>
+            <column name="date_start" type="DATE"/>
+            <column name="date_end" type="DATE"/>
+            <column name="status" type="BIGINT"/>
+        </createTable>
+        <createTable tableName="reliz">
+            <column name="id" type="BIGINT">
+                <constraints nullable="false" primaryKey="true" primaryKeyName="pk_reliz"/>
+            </column>
+            <column name="caption" type="VARCHAR(50)"/>
+            <column name="description" type="VARCHAR(120)"/>
+            <column name="date_start" type="DATE"/>
+            <column name="date_end" type="DATE"/>
+        </createTable>
+        <createTable tableName="role">
+            <column name="id" type="BIGINT">
+                <constraints nullable="false" primaryKey="true" primaryKeyName="pk_role"/>
+            </column>
+            <column name="role" type="VARCHAR(10)"/>
+        </createTable>
+        <createTable tableName="status">
+            <column autoIncrement="true" name="id" type="BIGINT">
+                <constraints nullable="false" primaryKey="true" primaryKeyName="pk_status"/>
+            </column>
+            <column name="caption" type="VARCHAR(120)"/>
+        </createTable>
+        <createTable tableName="task">
+            <column name="id" type="BIGINT">
+                <constraints nullable="false" primaryKey="true" primaryKeyName="pk_task"/>
+            </column>
+            <column name="title" type="VARCHAR(30)"/>
+            <column name="descriptoin" type="VARCHAR(120)"/>
+            <column name="autor" type="BIGINT"/>
+            <column name="executor" type="BIGINT"/>
+            <column name="history" type="BIGINT"/>
+            <column name="relizes" type="BIGINT"/>
+            <column name="project" type="BIGINT"/>
+            <column name="status" type="BIGINT"/>
+        </createTable>
+        <createTable tableName="users">
+            <column name="id" type="BIGINT">
+                <constraints nullable="false" primaryKey="true" primaryKeyName="pk_users"/>
+            </column>
+            <column name="name" type="VARCHAR(30)"/>
+            <column name="second_name" type="VARCHAR(30)"/>
+            <column name="role" type="BIGINT"/>
+        </createTable>
+        <addForeignKeyConstraint baseColumnNames="autor" baseTableName="project" constraintName="FK_PROJECT_ON_AUTOR"
+                                 referencedColumnNames="id" referencedTableName="users"/>
+        <addForeignKeyConstraint baseColumnNames="status" baseTableName="project" constraintName="FK_PROJECT_ON_STATUS"
+                                 referencedColumnNames="id" referencedTableName="status"/>
+        <addForeignKeyConstraint baseColumnNames="autor" baseTableName="task" constraintName="FK_TASK_ON_AUTOR"
+                                 referencedColumnNames="id" referencedTableName="users"/>
+        <addForeignKeyConstraint baseColumnNames="executor" baseTableName="task" constraintName="FK_TASK_ON_EXECUTOR"
+                                 referencedColumnNames="id" referencedTableName="users"/>
+        <addForeignKeyConstraint baseColumnNames="history" baseTableName="task" constraintName="FK_TASK_ON_HISTORY"
+                                 referencedColumnNames="id" referencedTableName="history"/>
+        <addForeignKeyConstraint baseColumnNames="project" baseTableName="task" constraintName="FK_TASK_ON_PROJECT"
+                                 referencedColumnNames="id" referencedTableName="project"/>
+        <addForeignKeyConstraint baseColumnNames="relizes" baseTableName="task" constraintName="FK_TASK_ON_RELIZES"
+                                 referencedColumnNames="id" referencedTableName="reliz"/>
+        <addForeignKeyConstraint baseColumnNames="status" baseTableName="task" constraintName="FK_TASK_ON_STATUS"
+                                 referencedColumnNames="id" referencedTableName="status"/>
+        <addForeignKeyConstraint baseColumnNames="role" baseTableName="users" constraintName="FK_USERS_ON_ROLE"
+                                 referencedColumnNames="id" referencedTableName="role"/>
+    </changeSet>
+</databaseChangeLog>
